local_llm_end_embedding_models/requirements.txt
streamlit
transformers
torch
faiss-cpu
PyPDF2
tqdm
python-dotenv
sentencepiece
accelerate
